refactor: make about modal component
